Refactor RegisterCatlogModel and remove useEffect
